Align the ready predicate method with named scope
Prior to this change, the `ready?` method returned truthiness under a
different set of conditions than the `ready` scope.

This update removes the check on playable formats length, in favor of
checking only the same conditions as the `ready` scope – mainly if the
video has a `zencoder_job_state` of either “playback_ready” or
“finished”. Expecting these two methods to align seems like a
reasonable expectation for models that include the mixin.